desktop_pipeline/requirements.txt
# ============================================
# Core ML/DL Frameworks
# ============================================
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0

# ============================================
# Computer Vision
# ============================================
timm>=0.9.12
pillow>=10.1.0
opencv-python>=4.8.1

# ============================================
# Scientific Computing
# ============================================
numpy<2
scipy>=1.11.4
scikit-learn>=1.3.2

# ============================================
# Data Validation
# ============================================
pydantic>=2.5.0

# ============================================
# Edge Deployment - BiomedCLIP
# ============================================
onnx>=1.14.0
onnxruntime>=1.16.0

# ============================================
# Edge Deployment - MedGemma
# ============================================
llama-cpp-python>=0.2.50
gguf>=0.6.0

# ============================================
# Model Loading & Quantization
# ============================================
safetensors>=0.4.0
bitsandbytes>=0.41.3
sentencepiece>=0.1.99

# ============================================
# Utilities
# ============================================
tqdm>=4.66.1
python-dotenv>=1.0.0

# ============================================
# Development
# ============================================
pytest>=7.4.3
black>=23.12.1
